+++ CHANGELOG.md
@@ -43,6 +43,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Network utilities to preformed cached downloads with hash validation.
 - Development service got a new command, which can retrieve an argument passed
   to setuptools `setup` function within a `setup.py` file.
+- Subflows can now get inputs forwarded from parent flows.
+  - `forward` parameter in `DataFlow`
+  - `register_subflow` in `MemoryOrchestratorContext`
 ### Changed
 - All instances of `src_url` changed to `key`.
 - `readonly` parameter in source config is now changed to `readwrite`.